requirements.txt
# Core Agno framework
# agno

# Google Gemini API packages (both needed)
google-generativeai
google-genai

# Required dependencies
sqlalchemy
pydantic
rich
mysql-connector-python

# Environment variables
python-dotenv